Move test utils for reuse.
