+++ tests/integration/_lib/utils.js
@@ -128,5 +128,6 @@ module.exports = {
   copyFixture,
   crawl,
   scrape,
-  validateResults
+  validateResults,
+  testCache
 }
