+++ tests/integration/events/crawler/paginated-json-test.js
@@ -1,9 +1,9 @@
 process.env.NODE_ENV = 'testing'
 
-const utils = require('../utils.js')
+const utils =  require('../../_lib/utils.js')
 const test = require('tape')
 const path = require('path')
-const testCache = require('../../_lib/testcache.js')
+const testCache = utils.testCache
 
 const firstPage = {
   records: [
